Add files via upload
The aim of this project is to develop a user-friendly web application that can assist users in finding the meaning, synonyms, and antonyms of words with ease. In today's world, where the internet is the primary source of information, a web-based dictionary application like this can be an indispensable tool for students, professionals, and anyone who needs to find the meaning of a word quickly. The motivation for this project stems from the fact thatthere is a growing need for efficient and reliable web- based dictionary applications. While there are several dictionary applications available online, they either lack certain features or are not user-friendly. This project aims to fill the research gap by developing a web-based dictionary application that is both efficient and user-friendly